backend/requirements.txt
# ===========================
# ZAIRE Healthcare - Backend
# ===========================

# Framework principal
Django==5.1.6
djangorestframework==3.15.2

# Autenticación JWT
djangorestframework-simplejwt==5.4.0
PyJWT==2.10.1

# Base de datos SQL Server
mssql-django==1.5
pyodbc==5.2.0

# CORS (permitir conexión desde React Native)
django-cors-headers==4.7.0

# Variables de entorno
python-decouple==3.8

# Generación de PDF
reportlab==4.4.0

# Inteligencia Artificial
scikit-learn==1.6.1
numpy==2.1.3
pandas==2.2.3
joblib==1.4.2

# Servidor de producción
gunicorn==23.0.0

# Utilidades
django-filter==24.3

# Desarrollo
black==25.1.0
flake8==7.1.2
